rename: base 패키지에 model 하위 패키지 생성
